@@ -18,13 +18,13 @@
     ],
     "require": {
         "php": "^7.1",
-        "symfony/dependency-injection": "^2.8 || ^3.0",
-        "symfony/form": "^2.8 || ^3.0"
+        "symfony/dependency-injection": "^2.8 || ^3.2",
+        "symfony/form": "^2.8 || ^3.2"
     },
     "require-dev": {
         "doctrine/orm": "^2.4",
         "sllh/php-cs-fixer-styleci-bridge": "^2.0",
-        "symfony/phpunit-bridge": "^2.8 || ^3.0"
+        "symfony/phpunit-bridge": "^3.3"
     },
     "autoload": {
         "psr-4": { "Sonata\\DatagridBundle\\": "src/" }
