@@ -69,7 +69,7 @@ abstract class BaseProxyQuery implements ProxyQueryInterface
      */
     public function __call($name, $args)
     {
-        return call_user_func_array(array($this->queryBuilder, $name), $args);
+        return call_user_func_array([$this->queryBuilder, $name], $args);
     }
 
     /**
